feat: Add ui game layout
