+++ android/app/src/main/java/com/threego/MainApplication.java
@@ -5,6 +5,9 @@ import android.content.Context;
 import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
 import com.github.wumke.RNExitApp.RNExitAppPackage;
+import com.github.wumke.RNExitApp.RNExitAppPackage;
+import com.github.wumke.RNExitApp.RNExitAppPackage;
+import com.github.wumke.RNExitApp.RNExitAppPackage;
 import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
