+++ src/features/game1/index.js
@@ -1,0 +1,4 @@
+import Game1Container from "./container/Game1Container";
+import Game1 from "./presenter/Game1";
+
+export { Game1Container, Game1 };
