--- src/features/game1/presenter/Game.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Game = () => {
-  return <div />;
-};
-
-export default Game;
