+++ src/features/main/index.js
@@ -1,3 +1,4 @@
 import MainContainer from "./container/MainContainer";
+import Main from "./presenter/Main";
 
-export default MainContainer;
+export { MainContainer, Main };
